+++ fazic_sdl/src/main.rs
@@ -1,114 +1,22 @@
-//#![feature(alloc_system)]
-//extern crate alloc_system;
-
-use std::fs;
-use std::fs::File;
-use std::io::prelude::*;
-use std::path::Path;
-
 #[macro_use]
 extern crate log;
-extern crate simple_logger;
-
+extern crate fazic;
 extern crate rand;
 extern crate sdl2;
+extern crate simple_logger;
 
-extern crate fazic;
-
-use fazic::config::*;
-use fazic::file_system::FileSystem;
-use sdl2::event::Event;
-use sdl2::keyboard::*;
-use sdl2::pixels::PixelFormatEnum;
-
-use std::process;
-
-const SCALE: f32 = 2.0;
-const WIDTH: u32 = SCREEN_WIDTH as u32 * SCALE as u32;
-const HEIGHT: u32 = SCREEN_HEIGHT as u32 * SCALE as u32;
-const FAZIC_FS: &str = "../fazic_fs";
-const RGB_WIDTH: usize = SCREEN_WIDTH as usize * 3;
-
-pub struct DiskFileSystem {}
-
-impl DiskFileSystem {
-    pub fn new() -> DiskFileSystem {
-        DiskFileSystem {}
-    }
-}
-
-impl FileSystem for DiskFileSystem {
-    fn dir(&self) -> Result<Vec<String>, String> {
-        let with_path = format!("{}/", FAZIC_FS);
-        let paths = fs::read_dir(&with_path).unwrap();
-
-        let names = paths
-            .map(|entry| {
-                let entry = entry.unwrap();
-                let entry_path = entry.path();
-                let file_name = entry_path.file_name().unwrap();
-                let file_name_as_str = file_name.to_str().unwrap();
-                let file_name_as_string = String::from(file_name_as_str);
-                file_name_as_string
-            })
-            .collect::<Vec<String>>();
-        Ok(names)
-    }
-
-    fn load(&self, filename: &str) -> Result<String, String> {
-        let with_path = format!("{}/{}", FAZIC_FS, filename);
-        let path = Path::new(&with_path);
-        let mut result = String::new();
-
-        match File::open(path) {
-            Ok(mut file) => match file.read_to_string(&mut result) {
-                Ok(_) => Ok(result),
-                _ => Err("NOT FOUND".to_string()),
-            },
-            _ => Err("NOT_FOUND".to_string()),
-        }
-    }
-
-    fn save(&mut self, filename: &str, program: String) -> Result<(), String> {
-        let with_path = format!("{}/{}", FAZIC_FS, filename);
-        let path = Path::new(&with_path);
-
-        match File::create(path) {
-            Ok(mut file) => match file.write_all(program.as_bytes()) {
-                Ok(_) => Ok(()),
-                _ => Err("NOT SAVED".to_string()),
-            },
-            _ => Err("NOT SAVED".to_string()),
-        }
-    }
-}
+mod events;
+mod file_system;
+mod window;
 
 pub fn main() {
     #[cfg(debug_assertions)]
     simple_logger::SimpleLogger::new().init().unwrap();
 
     let ctx = sdl2::init().unwrap();
-    let video_ctx = ctx.video().unwrap();
 
-    let window = video_ctx
-        .window("fazic", WIDTH, HEIGHT)
-        .position_centered()
-        .opengl()
-        .build()
-        .unwrap();
-
-    let mut canvas = window.into_canvas().build().unwrap();
-    let texture_creator = canvas.texture_creator();
-
-    let mut texture = texture_creator
-        .create_texture_streaming(
-            PixelFormatEnum::RGB24,
-            SCREEN_WIDTH as u32,
-            SCREEN_HEIGHT as u32,
-        )
-        .unwrap();
-
-    let mut events = ctx.event_pump().unwrap();
+    let mut window = window::Window::new(&ctx);
+    let mut events = events::Events::new(&ctx);
 
     let timer = ctx.timer().unwrap();
 
@@ -119,54 +27,13 @@ pub fn main() {
 
     let mut fazic = fazic::Fazic::new();
 
-    fazic.set_file_system(Box::new(DiskFileSystem::new()));
+    fazic.set_file_system(Box::new(file_system::DiskFileSystem::new()));
 
     loop {
         let main_loop_time = timer.ticks();
 
         if fps % 5 == 0 {
-            for event in events.poll_iter() {
-                match event {
-                    Event::Quit { .. } => process::exit(1),
-                    Event::KeyDown {
-                        keycode: Some(Keycode::Escape),
-                        ..
-                    } => fazic.stop_key(),
-                    Event::KeyDown {
-                        keycode: Some(key),
-                        keymod: Mod::LGUIMOD,
-                        ..
-                    }
-                    | Event::KeyDown {
-                        keycode: Some(key),
-                        keymod: Mod::RGUIMOD,
-                        ..
-                    } => match key {
-                        Keycode::Num1 => fazic.set_current_text_color(0),
-                        Keycode::Num2 => fazic.set_current_text_color(1),
-                        Keycode::Num3 => fazic.set_current_text_color(2),
-                        Keycode::Num4 => fazic.set_current_text_color(3),
-                        Keycode::Num5 => fazic.set_current_text_color(4),
-                        Keycode::Num6 => fazic.set_current_text_color(5),
-                        Keycode::Num7 => fazic.set_current_text_color(6),
-                        Keycode::Num8 => fazic.set_current_text_color(7),
-                        _ => (),
-                    },
-                    Event::KeyDown {
-                        keycode: Some(key), ..
-                    } => match key {
-                        Keycode::Left => fazic.left_key(),
-                        Keycode::Right => fazic.right_key(),
-                        Keycode::Up => fazic.up_key(),
-                        Keycode::Down => fazic.down_key(),
-                        Keycode::Backspace => fazic.backspace_key(),
-                        Keycode::Return => fazic.enter_key(),
-                        _ => (),
-                    },
-                    Event::TextInput { text: string, .. } => fazic.insert_string(string),
-                    _ => (),
-                }
-            }
+            events.process(&mut fazic);
         }
 
         if fps == 1 || fps == 30 {
@@ -174,13 +41,9 @@ pub fn main() {
         }
 
         if fazic.need_to_redraw() {
-            texture
-                .update(None, fazic.get_rgb_pixels(), RGB_WIDTH)
-                .unwrap();
-
-            let _ = canvas.copy(&texture, None, None);
-            canvas.present();
+            window.update(fazic.get_rgb_pixels());
         }
+
         if timer.ticks() - fps_last_time > 1000 {
             debug!("FPS: {}", fps);
             fps_last_time = timer.ticks();
